Add files via upload
It is my first task to redesign the e-commerce site of Cakelink.  In this project, I conducted a heuristic analysis of the website and based on that I found the problems and painpoints while using the website and I identified areas where improvements are needed. I create a user journey map, user flow diagram, mid-fidelity wireframe, design system, high-fidelity wireframe, and A/B testing. My goal is to make a website user-friendly and visually appealing interface and enhance the overall user experience with the website.